Change representation of selected tag
In the note editor we can select a tag before deleting it. This is a
confirmation mechanism to prevent accidental tag deletion.

Previously this "selected" tag was being stored as a reference to the
position in the tag list holding the specified tag. This system was
complicated and prone to error. In a previous commit
(c841e7bcffa6eefff96c7d08082fd781c2e02f91) I added logic to persist the
right selection after updates came in from outside sources.

However, the tags are specified in totality by the same string that
displays when showing the tag and is a sufficient reference in the code.

This patch stores the selected tag by its own name. By doing this we
obviate the need to rearrange references and shuffle around and it also
prevents storing the "selected" note index as `-1`, which is an
impossible index into the list.