Add export-to-zip functionality
This patch adds the ability to export the notes as a .zip in the same
way that this behavior works on app.simplenote.com

First we generate the JSON export and then process that data to build
the zip archive. The **JSZip** library does the hard work.

I have chosen to retain the naming of the notes to match the web app.
Changing that becomes rather trivial at this point however.