Introduce `react-virtualized` to fix performance issues in note list
@see https://github.com/bvaughn/react-virtualized

The note list has been a very slow component disproportionately slowing
down the rest of the app. There were a few reasons for this:

 - Every note was being rendered in the note list on every render. This
   meant that each note was processed rendered even if it not actually
   visible to the user.

 - Updating the search field quickly sorted the list exacerbating the
   issues with re-rendering all of the notes. Deferring to re-render, on
   the other hand, would obscure the search results.

`react-virtualized` only renders the notes in the note list which would
be visible and it automatically handles things like overscroll,
scrolling in general, and deciding how to update the list.

One challenge with Simplenote is that the rows in the note list have
varying heights depending on the content of the notes themselves. In
order to make the virtual list work properly, however, we must provide
the heights of each row.

In this patch we are supplying the height of each row by estimating the
number of lines of text in the note excerpt. This calculation is
**extremely fragile** and depends on many other tuned constants such as
font size, line height, and note list width. That means **this will
likely be a source of bugs and regressions** moving forward and **we
need to pay special attention to the row height calculation**.

In several places the behavior of the virtual list is not intuitive and
therefore there are places in the code which trigger recalculating the
row heights. This is necessary because the list is a very DOM-stateful
thing and does not synchronize normally with React as it handles prop
changes.